Yahia

+++ src/CanOperation.java
@@ -80,7 +80,7 @@ public interface CanOperation {
     void removePresentation(Movie m, Presentation p);
     
     void addComment(Customer customer, Movie movie, String Comment);
-    void addRate(Movie movie, int rate);
+    void addRate(Movie movie, float rate);
 
     Presentation getPresentation(Movie m, Date date, int hall_number);
     Ticket[][] getTickets(Presentation p);
